Move properties.js
