+++ src/properties.js
